docs: align cookbook entry with merged gl-sdk gitbook page (gl-sdk#5212)
