+++ gen-ai/how-to-guides/build_end_to_end_rag_pipeline/013_synthesize_responses_from_multiple_retrievers/pipeline.py
@@ -1,7 +1,7 @@
 """Example script to build and run a multi-retriever RAG pipeline.
 
 References:
-    [1] https://gdplabs.gitbook.io/sdk/how-to-guides/build-end-to-end-rag-pipeline/synthesize-responses-from-multiple-retrievers
+    [1] https://gdplabs.gitbook.io/sdk/gen-ai-sdk/guides/build-end-to-end-rag-pipeline/synthesize-responses-from-multiple-retrievers
 """
 
 import asyncio
@@ -10,16 +10,17 @@ from typing import Any, NotRequired, TypedDict
 
 from dotenv import load_dotenv
 from gllm_core.schema import Chunk
+from gllm_datastore.data_store import ChromaDataStore
+from gllm_datastore.data_store.chroma.data_store import ChromaClientType
 from gllm_generation.repacker import Repacker
 from gllm_generation.response_synthesizer import ResponseSynthesizer
+from gllm_inference.em_invoker.openai_em_invoker import OpenAIEMInvoker
 from gllm_inference.request_processor import build_lm_request_processor
 from gllm_pipeline.pipeline import Pipeline
 from gllm_pipeline.steps import parallel, step, transform
 from gllm_pipeline.types import Val
 from gllm_retrieval.chunk_processor import DedupeChunkProcessor
 from gllm_retrieval.retriever import SmartSearchWebRetriever, VectorRetriever
-
-from indexer import create_internal_retriever, ingest_internal_data
 
 load_dotenv()
 
@@ -43,6 +44,64 @@ class MultiRetrieverResearchState(TypedDict):
     response: NotRequired[str]
 
 
+def create_internal_retriever() -> VectorRetriever:
+    embedding_model = OpenAIEMInvoker(
+        model_name=os.environ["EMBEDDING_MODEL"],
+    )
+
+    data_store = ChromaDataStore(
+        collection_name="internal_research_docs",
+        client_type=ChromaClientType.PERSISTENT,
+        persist_directory="data/chroma",
+    ).with_vector(em_invoker=embedding_model)
+
+    return VectorRetriever(data_store=data_store)
+
+
+async def ingest_internal_data(vector_retriever: VectorRetriever) -> None:
+    """Index sample chunks so this example is runnable from a clean project.
+
+    In production, run ingestion separately; see index-your-data-with-vector-data-store.md.
+    """
+    chunks = [
+        Chunk(
+            id="internal-pipeline-overview",
+            content="GL SDK pipelines orchestrate steps, parallel branches, and response synthesis.",
+            metadata={"source": "internal-docs"},
+        ),
+        Chunk(
+            id="internal-vector-search",
+            content="VectorRetriever searches ChromaDB for internal chunks that match a user query.",
+            metadata={"source": "internal-docs"},
+        ),
+        Chunk(
+            id="internal-response-synthesis",
+            content="ResponseSynthesizer generates an answer from retrieved context chunks.",
+            metadata={"source": "internal-docs"},
+        ),
+    ]
+    await vector_retriever.data_store.vector.create(chunks)
+
+
+async def create_web_retriever() -> SmartSearchWebRetriever:
+    return await SmartSearchWebRetriever.create(
+        base_url=os.environ["SMART_SEARCH_BASE_URL"],
+        token=os.environ["SMART_SEARCH_TOKEN"],
+    )
+
+
+def create_response_synthesizer() -> ResponseSynthesizer:
+    lm_request_processor = build_lm_request_processor(
+        model_id=os.environ["LANGUAGE_MODEL"],
+        system_template=SYSTEM_PROMPT,
+        user_template=USER_PROMPT,
+    )
+    return ResponseSynthesizer.stuff(
+        lm_request_processor=lm_request_processor,
+        chunks_repacker=Repacker(mode="chunk"),
+    )
+
+
 def tag_chunks(chunks: list[Chunk], source_type: str) -> list[Chunk]:
     return [
         chunk.model_copy(update={"metadata": {**chunk.metadata, "source_type": source_type}}, deep=True)
@@ -60,74 +119,47 @@ def limit_chunks(state: dict[str, Any]) -> list[Chunk]:
     return state["chunks"][: state["merged_top_k"]]
 
 
-async def create_web_retriever() -> SmartSearchWebRetriever:
-    return await SmartSearchWebRetriever.create(
-        base_url=os.environ["SMART_SEARCH_BASE_URL"],
-        token=os.environ["SMART_SEARCH_TOKEN"],
-    )
-
-
-def create_response_synthesizer() -> ResponseSynthesizer:
-    lm_request_processor = build_lm_request_processor(
-        model_id=os.environ["LANGUAGE_MODEL"],
-        credentials=os.environ["OPENAI_API_KEY"],
-        system_template=SYSTEM_PROMPT,
-        user_template=USER_PROMPT,
-    )
-    return ResponseSynthesizer.stuff(
-        lm_request_processor=lm_request_processor,
-        chunks_repacker=Repacker(mode="chunk"),
-    )
-
-
 def build_pipeline(
     web_retriever: SmartSearchWebRetriever,
     vector_retriever: VectorRetriever,
     response_synthesizer: ResponseSynthesizer,
 ) -> Pipeline:
     return Pipeline(
-        [
+        steps=[
             parallel(
                 branches={
                     "web": step(
                         web_retriever,
                         input_map={"query": "query", "top_k": Val(5)},
                         output_state="web_chunks",
-                        name="web_retriever",
                     ),
                     "vector": step(
                         vector_retriever,
                         input_map={"query": "query", "top_k": Val(5)},
                         output_state="vector_chunks",
-                        name="vector_retriever",
                     ),
                 },
                 input_states=["query"],
-                name="parallel_retrieval",
             ),
             transform(
                 combine_retrieved_chunks,
                 input_states=["web_chunks", "vector_chunks"],
                 output_state="chunks",
-                name="combine_retrieved_chunks",
             ),
             step(
                 DedupeChunkProcessor(),
                 input_map={"chunks": "chunks"},
                 output_state="chunks",
-                name="dedupe_chunks",
             ),
             transform(
                 limit_chunks,
                 input_states=["chunks", "merged_top_k"],
                 output_state="chunks",
-                name="limit_chunks",
             ),
             step(
                 response_synthesizer,
                 input_map={"query": "query", "chunks": "chunks"},
                 output_state="response",
-                name="response_synthesizer",
             ),
         ],
         state_type=MultiRetrieverResearchState,
